art body is one component w conditional render
