--- src/components/art-page/Body/After/PromptBody.css
@@ -1,19 +1,0 @@
-.pagebody {
-    max-width: 1200px;
-    margin: 0 auto;
-}
-
-.gridcontainer {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-
-}
-
-.gridspace {
-    border: 4px solid black;
-    background-color: aqua;
-    /* max-width: 400px; */
-    min-height: 300px;
-    margin: 10px;
-}
